requirements.txt
PyYAML==3.12
python-dateutil==2.6.1
addict==2.1.3
airtest==1.0.25
schedule==0.6.0